+++ packages/http/generated-defs/TypeSpec.Http.Private.ts
@@ -1,8 +1,9 @@
 import type { DecoratorContext, Model, Type } from "@typespec/compiler";
 
-interface HttpPartOptions {
+export interface HttpPartOptions {
   readonly name?: string;
 }
+
 export type PlainDataDecorator = (context: DecoratorContext, target: Model) => void;
 
 export type HttpFileDecorator = (context: DecoratorContext, target: Model) => void;
